CLOUDIFY-1018 removed unnecessary cast
git-svn-id: svn://pc-lab14/SVN/cloudify/trunk/cloudify@125535 eb64e737-3616-4df0-8941-5ee2ae88103d

+++ USM/src/main/java/org/cloudifysource/usm/USMLifecycleBean.java
@@ -218,7 +218,7 @@ public class USMLifecycleBean implements ClusterInfoAware {
 			long eventExecDuration = System.currentTimeMillis() - eventStartTime;
 			String durationAsString = DurationFormatUtils.formatDuration(eventExecDuration, "s.S");
 			float formattedDurationAsLong = Float.parseFloat(durationAsString);
-			String formattedDurationAsString = String.format("%.1f", (float) formattedDurationAsLong);
+			String formattedDurationAsString = String.format("%.1f", formattedDurationAsLong);
 			eventLogger.info(eventPrefix + event + CloudifyConstants.USM_EVENT_EXEC_SUCCESSFULLY 
 					+ ", duration: " + formattedDurationAsString + " seconds");
 		}
